3.4.3 test nowy uzytkownik moze dac like do postu
