+++ blogApiTests/src/test/java/edu/iis/mto/blog/rest/test/AddLikePostTest.java
@@ -1,0 +1,20 @@
+package edu.iis.mto.blog.rest.test;
+
+import com.jayway.restassured.RestAssured;
+import com.jayway.restassured.http.ContentType;
+import org.apache.http.HttpStatus;
+import org.json.JSONObject;
+import org.junit.Test;
+
+/**
+ * Created by Patryk on 2017-05-23.
+ */
+public class AddLikePostTest {
+    @Test
+    public void userWithStatusNewShouldNotLikePost() {
+        JSONObject jsonObj = new JSONObject().put("entry", "post");
+        RestAssured.given().accept(ContentType.JSON).header("Content-Type", "application/json;charset=UTF-8")
+                .body(jsonObj.toString()).expect().log().all().statusCode(HttpStatus.SC_BAD_REQUEST).when()
+                .post("/blog/user/2/like/1");
+    }
+}
