+++ blogApiTests/src/test/java/edu/iis/mto/blog/rest/test/AddPostTest.java
@@ -1,0 +1,27 @@
+package edu.iis.mto.blog.rest.test;
+
+import com.jayway.restassured.RestAssured;
+import com.jayway.restassured.http.ContentType;
+import org.apache.http.HttpStatus;
+import org.json.JSONObject;
+import org.junit.Test;
+
+/**
+ * Created by Patryk on 2017-05-23.
+ */
+public class AddPostTest {
+    @Test
+    public void userWithStatusConfirmedShouldAddPost() {
+        JSONObject jsonObj = new JSONObject().put("entry", "post");
+        RestAssured.given().accept(ContentType.JSON).header("Content-Type", "application/json;charset=UTF-8")
+                .body(jsonObj.toString()).expect().log().all().statusCode(HttpStatus.SC_CREATED).when()
+                .post("/blog/user/1/post");
+    }
+    @Test
+    public void userWithStatusNewShouldNotAddPost() {
+        JSONObject jsonObj = new JSONObject().put("entry", "post");
+        RestAssured.given().accept(ContentType.JSON).header("Content-Type", "application/json;charset=UTF-8")
+                .body(jsonObj.toString()).expect().log().all().statusCode(HttpStatus.SC_BAD_REQUEST).when()
+                .post("/blog/user/2/post");
+    }
+}
